requirements.txt
alembic==1.8.1
greenlet==2.0.1
importlib-metadata==5.0.0
importlib-resources==5.10.0
Mako==1.2.3
MarkupSafe==2.1.1
numpy==1.23.4
pandas==1.5.1
psycopg2==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.44
zipp==3.10.0
